create provider, fix styles for storybook
